Move files to a new "profiling" directory to better distinguish from census statistics
