+++ BUILD
@@ -186,12 +186,12 @@ cc_library(
     "src/core/json/json_common.h",
     "src/core/json/json_reader.h",
     "src/core/json/json_writer.h",
+    "src/core/profiling/timers.h",
     "src/core/statistics/census_interface.h",
     "src/core/statistics/census_log.h",
     "src/core/statistics/census_rpc_stats.h",
     "src/core/statistics/census_tracing.h",
     "src/core/statistics/hash_table.h",
-    "src/core/statistics/timers.h",
     "src/core/statistics/window_stats.h",
     "src/core/surface/byte_buffer_queue.h",
     "src/core/surface/call.h",
@@ -298,12 +298,12 @@ cc_library(
     "src/core/json/json_reader.c",
     "src/core/json/json_string.c",
     "src/core/json/json_writer.c",
+    "src/core/profiling/timers.c",
     "src/core/statistics/census_init.c",
     "src/core/statistics/census_log.c",
     "src/core/statistics/census_rpc_stats.c",
     "src/core/statistics/census_tracing.c",
     "src/core/statistics/hash_table.c",
-    "src/core/statistics/timers.c",
     "src/core/statistics/window_stats.c",
     "src/core/surface/byte_buffer.c",
     "src/core/surface/byte_buffer_queue.c",
@@ -415,12 +415,12 @@ cc_library(
     "src/core/json/json_common.h",
     "src/core/json/json_reader.h",
     "src/core/json/json_writer.h",
+    "src/core/profiling/timers.h",
     "src/core/statistics/census_interface.h",
     "src/core/statistics/census_log.h",
     "src/core/statistics/census_rpc_stats.h",
     "src/core/statistics/census_tracing.h",
     "src/core/statistics/hash_table.h",
-    "src/core/statistics/timers.h",
     "src/core/statistics/window_stats.h",
     "src/core/surface/byte_buffer_queue.h",
     "src/core/surface/call.h",
@@ -507,12 +507,12 @@ cc_library(
     "src/core/json/json_reader.c",
     "src/core/json/json_string.c",
     "src/core/json/json_writer.c",
+    "src/core/profiling/timers.c",
     "src/core/statistics/census_init.c",
     "src/core/statistics/census_log.c",
     "src/core/statistics/census_rpc_stats.c",
     "src/core/statistics/census_tracing.c",
     "src/core/statistics/hash_table.c",
-    "src/core/statistics/timers.c",
     "src/core/statistics/window_stats.c",
     "src/core/surface/byte_buffer.c",
     "src/core/surface/byte_buffer_queue.c",
