+++ tools/run_tests/tests.json
@@ -334,6 +334,11 @@
   {
     "flaky": false, 
     "language": "c", 
+    "name": "timers_test"
+  }, 
+  {
+    "flaky": false, 
+    "language": "c", 
     "name": "transport_metadata_test"
   }, 
   {
